Fix: POC-Preis auf natives float casten
POC wurde anders als VAL/VAH nicht auf einen nativen Python-float
gecastet und blieb ein numpy.float64. Sobald POC das nächstgelegene
Level war, propagierte sich das über distance_pct bis zu is_hit, das
dann als numpy.bool statt Python bool berechnet wurde – nicht
JSON-serialisierbar, wodurch der Report-Schritt im GitHub-Actions-Lauf
mit "Object of type bool is not JSON serializable" abbrach.

+++ src/volume_profile.py
@@ -49,7 +49,7 @@ def compute_volume_profile(df: pd.DataFrame, bins: int = 30, value_area_pct: flo
         raise ValueError("Kein Volumen für Volumenprofil vorhanden")
 
     poc_idx = int(np.argmax(bin_volume))
-    poc_price = (edges[poc_idx] + edges[poc_idx + 1]) / 2
+    poc_price = float((edges[poc_idx] + edges[poc_idx + 1]) / 2)
 
     # Value Area: von POC ausgehend jeweils die volumenstärkere Nachbar-Bin
     # dazunehmen, bis value_area_pct % des Gesamtvolumens erreicht sind.
